First eal cut of the PHA Tools.
Bring in all of the moving parts that're ongoing from RPE Tools
that really don't belong in the foundational core (Which will
stay as RPE Tools...) into a new repo.  It's not done-done
as one of the class objects just simply cowpiles and still
needs another couple of abstractions that let people frame
in multiple GPIO controls into a cleaner abstraction than
the one done by the libgpiod people that is either better
in performance or easier to follow.

(This crap was bad...poor docs...inadequate demos to explain
 themselves...and they re-did the simpler, cleaner API
 for the much more confusing and complicated arbitrary
 multi-line framework...)